Modified dictionary module to add newer functionality
Generically, a GET of the dictionary module's URL with the following request variables are supported

Required ( Omitting any of these results in a server error )
========
image=<value> --> if value=="y", a renderd image(png) of the word is retrieved. else meaning in Uniocode text will be retrieved.
text=<word>   --> word whose meaning is to be looked up
dict=<en-ml>  --> the string that denotes the dictionary to be used

Optional ( Supported only when image is set to "y", needs more testing)
========
imagewidth ( defaults to 300 )
imageheight ( defaults to 200 )
fontsize ( defaults to 12 )

eg:
http://<silpaserver>/Dictionary?image=y&text=help&dict=en-ml&imagewidth=500&imageheight=500&fontsize=16

The render module was modified to accept fontsize as an optional parameter as well.